backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
anthropic
beautifulsoup4
lxml
python-dotenv
